+++ apps/tergite-mss/tests/api/rest/test_devices.py
@@ -10,36 +10,121 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """Integration tests for the devices router"""
-from typing import Any, Dict
+from typing import Any, Dict, List, Optional
 
 import pytest
 
 from tests._utils.date_time import is_not_older_than
 from tests._utils.fixtures import load_json_fixture
 from tests._utils.mongodb import find_in_collection, insert_in_collection
-from tests._utils.records import get_record, order_by, pop_field
+from tests._utils.records import (
+    filter_by_equality,
+    get_record,
+    order_by_many,
+    pop_field,
+    with_incremental_timestamps,
+)
+from tests.api.rest.test_calibrations import _attach_str_ids
 
 _DEVICES_COLLECTION = "devices"
 _EXCLUDED_FIELDS = ["_id"]
 
 _DEVICE_LIST = load_json_fixture("device_list.json")
+_SKIP_LIMIT_SORT_PARAMS = [
+    (0, 1, ["-version", "updated_at"]),
+    (2, 4, None),
+    (1, 3, ["version", "created_at"]),
+    (None, 10, None),
+    (2, None, ["created_at"]),
+]
+_SEARCH_PARAMS = [
+    {"version": "2024.04.1"},
+    {"name": "Loke"},
+    {"version": "2023.06.0"},
+    {"version": "2023.06.0", "name": "Loke"},
+    {"version": "2023.06.0", "name": "Thor"},
+    {},
+]
+_PAGINATE_AND_SEARCH_PARAMS = [
+    (skip, limit, sort, search)
+    for skip, limit, sort in _SKIP_LIMIT_SORT_PARAMS
+    for search in _SEARCH_PARAMS
+]
 
 
-def test_read_devices(db, client, user_jwt_cookie):
-    """GET to /devices/ retrieves all devices"""
-    insert_in_collection(
-        database=db, collection_name=_DEVICES_COLLECTION, data=_DEVICE_LIST
+@pytest.mark.parametrize("skip, limit, sort, search", _PAGINATE_AND_SEARCH_PARAMS)
+def test_find_devices(
+    db,
+    client,
+    skip: Optional[int],
+    limit: Optional[int],
+    sort: Optional[List[str]],
+    search: dict,
+    freezer,
+):
+    """Get to /devices/?version=...&name=... can search for the devices that fulfill the given filters"""
+    raw_devices = with_incremental_timestamps(
+        _DEVICE_LIST, fields=("created_at", "updated_at")
     )
+    inserted_ids = insert_in_collection(
+        database=db, collection_name=_DEVICES_COLLECTION, data=raw_devices
+    )
+    raw_devices = _attach_str_ids(raw_devices, ids=inserted_ids, id_field="id")
+
+    query_string = "?"
+    slice_end = len(raw_devices)
+    slice_start = 0
+    sort_fields = [
+        "-created_at",
+    ]
+    if limit is not None:
+        query_string += f"limit={limit}&"
+        slice_end = limit
+    if skip is not None:
+        query_string += f"skip={skip}&"
+        slice_start = skip
+        slice_end += skip
+    if sort is not None:
+        sort_fields = sort
+        for sort_field in sort_fields:
+            query_string += f"sort={sort_field}&"
+
+    # Adding search
+    for key, value in search.items():
+        query_string += f"{key}={value}&"
 
     # using context manager to ensure on_startup runs
     with client as client:
-        response = client.get(f"/devices/", cookies=user_jwt_cookie)
-        got = order_by(response.json(), field="name")
-        pop_field(got, field="id")
-        expected = order_by(_DEVICE_LIST, field="name")
+        response = client.get(f"/devices/{query_string}")
+        got = response.json()
+
+        filtered_data = filter_by_equality(raw_devices, filters=search)
+        sorted_data = order_by_many(filtered_data, fields=sort_fields)
+        expected = {
+            "skip": slice_start,
+            "limit": limit,
+            "data": sorted_data[slice_start:slice_end],
+        }
 
         assert response.status_code == 200
         assert got == expected
+
+
+# def test_read_devices(db, client, user_jwt_cookie):
+#     """GET to /devices/ retrieves all devices"""
+#     insert_in_collection(
+#         database=db, collection_name=_DEVICES_COLLECTION, data=_DEVICE_LIST
+#     )
+#
+#     # using context manager to ensure on_startup runs
+#     with client as client:
+#         response = client.get(f"/devices/", cookies=user_jwt_cookie)
+#         got = order_by(response.json(), field="name")
+#         pop_field(got, field="id")
+#         expected = order_by(_DEVICE_LIST, field="name")
+#
+#         assert response.status_code == 200
+#         assert got == expected
 
 
 @pytest.mark.parametrize("name", [v["name"] for v in _DEVICE_LIST])